Move BlockREgionTransform class.
