+++ src/main/java/org/terasology/structureTemplates/components/BlockRegionTransformComponent.java
@@ -17,7 +17,7 @@ package org.terasology.structureTemplates.components;
 
 import org.terasology.entitySystem.Component;
 import org.terasology.math.geom.Vector3i;
-import org.terasology.structureTemplates.util.transform.BlockRegionTransform;
+import org.terasology.structureTemplates.util.BlockRegionTransform;
 
 /**
  * Stores the data to construct a {@link BlockRegionTransform}
