+++ src/main/java/org/terasology/structureTemplates/events/StructureSpawnStartedEvent.java
@@ -16,7 +16,7 @@
 package org.terasology.structureTemplates.events;
 
 import org.terasology.entitySystem.event.Event;
-import org.terasology.structureTemplates.util.transform.BlockRegionTransform;
+import org.terasology.structureTemplates.util.BlockRegionTransform;
 
 /**
  * The event gets sent to a structure template entity when a structure template spawning process started.
