+++ src/main/java/org/terasology/structureTemplates/internal/systems/StructureTemplateEditorServerSystem.java
@@ -66,7 +66,7 @@ import org.terasology.structureTemplates.internal.events.StopEditingProcessReque
 import org.terasology.structureTemplates.internal.events.StructureTemplateStringRequest;
 import org.terasology.structureTemplates.util.ListUtil;
 import org.terasology.structureTemplates.util.RegionMergeUtil;
-import org.terasology.structureTemplates.util.transform.BlockRegionTransform;
+import org.terasology.structureTemplates.util.BlockRegionTransform;
 import org.terasology.world.BlockEntityRegistry;
 import org.terasology.world.WorldProvider;
 import org.terasology.world.block.Block;
