+++ src/main/java/org/terasology/structureTemplates/util/BlockRegionTransform.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 MovingBlocks
+ * Copyright 2017 MovingBlocks
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.structureTemplates.util.transform;
+package org.terasology.structureTemplates.util;
 
 import org.terasology.math.Region3i;
 import org.terasology.math.Side;
