+++ src/test/java/org/terasology/structureTemplates/internal/systems/ScheduledStructureSpawnSystemTest.java
@@ -18,7 +18,7 @@ package org.terasology.structureTemplates.internal.systems;
 import org.junit.Test;
 import org.terasology.math.Side;
 import org.terasology.math.geom.Vector3i;
-import org.terasology.structureTemplates.util.transform.BlockRegionTransform;
+import org.terasology.structureTemplates.util.BlockRegionTransform;
 
 import static org.junit.Assert.assertEquals;
 import static org.terasology.structureTemplates.internal.systems.ScheduledStructureSpawnSystem.createTransformForIncomingConnectionPoint;
